Additional Island Info Popup
Added functionality for a modal-like box containing additional/complete
info on an island when you double click it. This works independently of
the leaflet library, and shows/hides an additional html division above
the map. This will need some tweaks in the future but it should do the
job for now